# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khalaf Olive Groves - Ancient Grove Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groves:
    get:
      operationId: listGroves
      tags: [groves]
      summary: List all groves
      responses:
        "200":
          description: List of groves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grove"
    post:
      operationId: createGrove
      tags: [groves]
      summary: Create a grove
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroveRequest"
      responses:
        "201":
          description: Created grove
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grove"
  /groves/{id}:
    get:
      operationId: getGrove
      tags: [groves]
      summary: Get grove by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grove
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grove"
    patch:
      operationId: updateGrove
      tags: [groves]
      summary: Update a grove
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroveRequest"
      responses:
        "200":
          description: Updated grove
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grove"
  /groves/{id}/summary:
    get:
      operationId: getGroveSummary
      tags: [groves]
      summary: Grove summary with tree/alert/task counts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grove summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroveSummary"

  # ─── Trees ───────────────────────────────────────────────────────────────────
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List logged field activities
      parameters:
        - in: query
          name: groveId
          schema: { type: integer }
        - in: query
          name: treeId
          schema: { type: integer }
        - in: query
          name: workerId
          schema: { type: integer }
        - in: query
          name: activityType
          schema: { type: string }
        - in: query
          name: fromDate
          schema: { type: string }
        - in: query
          name: toDate
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log a new activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get activity by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Phenology ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Grove:
      type: object
      properties:
        id:
          type: integer
        groveCode:
          type: string
        name:
          type: string
        boundaryGeojson:
          type: object
          nullable: true
        areaHa:
          type: number
          nullable: true
        centroidLat:
          type: number
          nullable: true
        centroidLon:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        heritageNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - groveCode
        - name
        - createdAt
        - updatedAt
    CreateGroveRequest:
      type: object
      properties:
        groveCode:
          type: string
        name:
          type: string
        boundaryGeojson:
          type: object
          nullable: true
        areaHa:
          type: number
          nullable: true
        centroidLat:
          type: number
          nullable: true
        centroidLon:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        heritageNotes:
          type: string
          nullable: true
      required:
        - groveCode
        - name
    UpdateGroveRequest:
      type: object
      properties:
        name:
          type: string
        notes:
          type: string
          nullable: true
        heritageNotes:
          type: string
          nullable: true
        areaHa:
          type: number
          nullable: true
        centroidLat:
          type: number
          nullable: true
        centroidLon:
          type: number
          nullable: true
        boundaryGeojson:
          type: object
          nullable: true
    GroveSummary:
      type: object
      properties:
        grove:
          $ref: "#/components/schemas/Grove"
        treeCount:
          type: integer
        ancientTreeCount:
          type: integer
        openAlertCount:
          type: integer
        openTaskCount:
          type: integer
        averageHealthIndex:
          type: number
          nullable: true
        harvestedTreesThisSeason:
          type: integer
        recentVisits:
          type: integer
      required:
        - grove
        - treeCount
        - ancientTreeCount
        - openAlertCount
        - openTaskCount
        - harvestedTreesThisSeason
        - recentVisits

    # ─── Tree ──────────────────────────────────────────────────────────────────
...
    Activity:
      type: object
      required: [id, workerId, groveId, activityType, performedAt, createdAt]
      properties:
        id: { type: integer }
        workerId: { type: integer }
        workerName: { type: string, nullable: true }
        groveId: { type: integer }
        groveName: { type: string, nullable: true }
        treeIds:
          type: array
          nullable: true
          items: { type: integer }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        taskId: { type: integer, nullable: true }
        activityType:
          type: string
          enum:
            - pruning_light
            - pruning_structural
            - pruning_heavy
            - mowing
            - terrace_repair
            - stone_drainage
            - cover_crop
            - tree_planting
            - tree_removal
            - inspection
            - cleaning
            - damage_assessment
            - other
        performedAt: { type: string, format: date-time }
        durationMinutes: { type: integer, nullable: true }
        treesAffectedCount: { type: integer, nullable: true }
        areaHectares: { type: number, nullable: true }
        materialsUsed: { type: string, nullable: true }
        gpsLat: { type: number, nullable: true }
        gpsLon: { type: number, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateActivityRequest:
      type: object
      required: [workerId, groveId, activityType]
      properties:
        workerId: { type: integer }
        groveId: { type: integer }
        treeIds:
          type: array
          nullable: true
          items: { type: integer }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        taskId: { type: integer, nullable: true }
        activityType:
          type: string
          enum:
            - pruning_light
            - pruning_structural
            - pruning_heavy
            - mowing
            - terrace_repair
            - stone_drainage
            - cover_crop
            - tree_planting
            - tree_removal
            - inspection
            - cleaning
            - damage_assessment
            - other
        performedAt: { type: string, format: date-time, nullable: true }
        durationMinutes: { type: integer, nullable: true }
        treesAffectedCount: { type: integer, nullable: true }
        areaHectares: { type: number, nullable: true }
        materialsUsed: { type: string, nullable: true }
        gpsLat: { type: number, nullable: true }
        gpsLon: { type: number, nullable: true }
        notes: { type: string, nullable: true }
    UpdateActivityRequest:
      type: object
      properties:
        groveId: { type: integer, nullable: true }
        treeIds:
          type: array
          nullable: true
          items: { type: integer }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        activityType: { type: string, nullable: true }
        performedAt: { type: string, format: date-time, nullable: true }
        durationMinutes: { type: integer, nullable: true }
        treesAffectedCount: { type: integer, nullable: true }
        areaHectares: { type: number, nullable: true }
        materialsUsed: { type: string, nullable: true }
        gpsLat: { type: number, nullable: true }
        gpsLon: { type: number, nullable: true }
        notes: { type: string, nullable: true }
...